+++ core/pkg/resourcehandler/interface.go
@@ -13,4 +13,5 @@ import (
 type IResourceHandler interface {
 	GetResources(context.Context, *cautils.OPASessionObj, *armotypes.PortalDesignator, opaprocessor.IJobProgressNotificationClient) (cautils.K8SResources, map[string]workloadinterface.IMetadata, cautils.KSResources, map[string]bool, error)
 	GetClusterAPIServerInfo(ctx context.Context) *version.Info
+	GetWorkloadParentKind(workloadinterface.IWorkload) string
 }
