+++ core/pkg/resourcehandler/k8sresources.go
@@ -74,7 +74,7 @@ func (k8sHandler *K8sResourceHandler) GetResources(ctx context.Context, sessionO
 	resourceToControl := make(map[string][]string)
 	// build resources map
 	// map resources based on framework required resources: map["/group/version/kind"][]<k8s workloads ids>
-	queryableResources, excludedRulesMap := getQueryableResourceMapFromPolicies(sessionObj.Policies, workload)
+	queryableResources, excludedRulesMap := getQueryableResourceMapFromPolicies(k8sHandler, sessionObj.Policies, workload)
 	ksResourceMap := setKSResourceMap(sessionObj.Policies, resourceToControl)
 
 	// get namespace and labels from designator (ignore cluster labels)
@@ -182,7 +182,7 @@ func (k8sHandler *K8sResourceHandler) findWorkloadToScan(workloadIdentifier *cau
 		return nil, err
 	}
 
-	result, err := k8sHandler.pullSingleResource(&gvr, workloadIdentifier.Namespace, nil, getNameSelector(workloadIdentifier.Name))
+	result, err := k8sHandler.pullSingleResource(&gvr, workloadIdentifier.Namespace, nil, getNameFieldSelector(workloadIdentifier.Name, "="))
 	if err != nil {
 		return nil, fmt.Errorf("failed to get resource %s, reason: %v", workloadIdentifier.String(), err)
 	}
@@ -349,6 +349,15 @@ func (k8sHandler *K8sResourceHandler) pullResources(queryableResources Queryable
 		}
 	}
 	return k8sResources, allResources, errs
+}
+
+func (k8sHandler *K8sResourceHandler) GetWorkloadParentKind(workload workloadinterface.IWorkload) string {
+	if workload == nil {
+		return ""
+	}
+
+	kind, _, _ := k8sHandler.k8s.CalculateWorkloadParentRecursive(workload)
+	return kind
 }
 
 func (k8sHandler *K8sResourceHandler) pullSingleResource(resource *schema.GroupVersionResource, namespace string, labels map[string]string, fields string) ([]unstructured.Unstructured, error) {
